Realized I could pull the spinning up the render loop to the display module.
