+++ cli/src/main.rs
@@ -1,18 +1,16 @@
 use std::{
     env,
-    sync::{
-        mpsc::{channel, Receiver, Sender},
-        Arc,
-    },
-    thread::{self, JoinHandle},
+    sync::{mpsc::Sender, Arc},
+    thread::JoinHandle,
 };
 
 use clap::{Parser, Subcommand};
 use clap_verbosity_flag::{log::LevelFilter, InfoLevel, Verbosity};
-use display::message::{send_message, DisplayMessage};
+use display::{
+    message::{send_message, DisplayMessage},
+    render::spin_up_renderer,
+};
 use sentry::ClientInitGuard;
-use superconsole::{Dimensions, SuperConsole};
-use terminal_size::{terminal_size, Height, Width};
 use tracing_subscriber::{filter::FilterFn, prelude::*};
 use trunk_analytics_cli::{
     context::gather_debug_props,
@@ -91,43 +89,11 @@ enum Commands {
     Validate(ValidateArgs),
 }
 
-fn output_render_loop(receiver: Receiver<DisplayMessage>) {
-    let mut superconsole = match SuperConsole::new() {
-        Some(console) => console,
-        None => {
-            tracing::warn!("Failed to create superconsole because of incompatible TTY");
-            let size = terminal_size();
-            if let Some((Width(width), Height(height))) = size {
-                SuperConsole::forced_new(Dimensions {
-                    width: width as usize,
-                    height: height as usize,
-                })
-            } else {
-                tracing::warn!("Falling back to default dimensions for superconsole");
-                // Fallback to a reasonable default size
-                // This is a fallback for when the terminal size cannot be determined
-                // or when running in a non-TTY environment.
-                tracing::warn!("Using default dimensions for superconsole");
-                tracing::warn!("This may not render correctly in some environments");
-                tracing::warn!("Please use a TTY compatible terminal for best results");
-                SuperConsole::forced_new(Dimensions {
-                    width: 143,
-                    height: 24,
-                })
-            }
-        }
-    };
-
-    while let Ok(message) = receiver.recv() {
-        message.render(&mut superconsole);
-    }
-}
-
-fn close_out_and_exit<T>(
+fn close_out_and_exit(
     exit_code: i32,
     guard: ClientInitGuard,
     render_sender: Sender<DisplayMessage>,
-    render_handle: JoinHandle<T>,
+    render_handle: JoinHandle<()>,
 ) -> ! {
     guard.flush(None);
     drop(render_sender);
@@ -159,10 +125,7 @@ fn main() -> anyhow::Result<()> {
                 command = cli.debug_props(),
                 "Trunk Flaky Test running command"
             );
-            let (render_sender, render_receiver) = channel::<DisplayMessage>();
-            let render_handle = thread::spawn(move || {
-                output_render_loop(render_receiver);
-            });
+            let (render_handle, render_sender) = spin_up_renderer();
 
             match run(cli, render_sender.clone()).await {
                 Ok(RunResult::Upload(run_result)) => {
