+++ display/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 anyhow = "1.0.44"
 superconsole = "0.2.0"
 tracing = "0.1.41"
+terminal_size = "0.4.2"
