+++ display/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod end_output;
 pub mod message;
+pub mod render;
